Revert code & Correct flow for Welcom Modal
