+++ src/components/Panels/ModalPanel.tsx
@@ -14,9 +14,6 @@ type Props = {
   className?: string;
   onRequestClose?: () => void;
   onAfterOpen?: () => void;
-  notCloseOnEsc?: boolean;
-  notCloseOnOverlayClick?: boolean;
-  containerClassName?: string;
 };
 
 const sizeClass = {
@@ -29,13 +26,10 @@ function ModalPanel(props: Props) {
   const {
     children,
     className,
-    containerClassName,
     fullHeight,
     header,
     noGap,
     noPadding,
-    notCloseOnEsc,
-    notCloseOnOverlayClick,
     onAfterOpen,
     onRequestClose,
     show,
@@ -51,19 +45,11 @@ function ModalPanel(props: Props) {
     <Modal
       isOpen={show}
       overlayClassName="fixed inset-0 bg-black/50 z-[80]"
-      className={cx('mt-10 flex h-full items-center justify-center', containerClassName)}
+      className="mt-10 flex h-full items-center justify-center"
       onAfterOpen={onAfterOpen}
       closeTimeoutMS={150}
-      shouldCloseOnEsc={notCloseOnEsc !== true}
-      shouldCloseOnOverlayClick={notCloseOnOverlayClick !== true}
     >
-      <div
-        className="flex size-full items-center justify-center"
-        onClick={() => {
-          if (notCloseOnOverlayClick) return;
-          onRequestClose?.();
-        }}
-      >
+      <div className="flex h-full w-full items-center justify-center" onClick={onRequestClose}>
         <div
           className={cx(
             'flex flex-col rounded-lg border border-panel-border bg-panel-background drop-shadow-lg',
