+++ src/pages/firstrun/Footer.tsx
@@ -1,8 +1,9 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import cx from 'classnames';
 
 import Button from '@/components/Input/Button';
+import WelcomeModal from '@/pages/firstrun/WelcomeModal';
 
 import type { TestStatusType } from '@/core/slices/firstrun';
 
@@ -25,6 +26,8 @@ function Footer(props: Props) {
     if (nextPage) navigate(`../${nextPage}`);
   };
 
+  const [showWelcomeModal, setShowWelcomeModal] = useState(false);
+
   const {
     finish,
     isFetching,
@@ -34,43 +37,47 @@ function Footer(props: Props) {
   } = props;
 
   return (
-    <div className="flex flex-col text-lg">
-      <div
-        className={cx([
-          'flex items-center mb-5',
-          status?.type === 'error' ? 'text-panel-text-danger' : 'text-panel-text-important',
-        ])}
-      >
-        {status?.text}
+    <>
+      <div className="flex flex-col text-lg">
+        <div
+          className={cx([
+            'flex items-center mb-5',
+            status?.type === 'error' ? 'text-panel-text-danger' : 'text-panel-text-important',
+          ])}
+        >
+          {status?.text}
+        </div>
+        <div className="flex justify-between font-semibold">
+          <Button onClick={() => navigate(-1)} buttonType="primary" className="mr-6 w-1/2 py-2" disabled={prevDisabled}>
+            Back
+          </Button>
+          {finish
+            ? (
+              <Button
+                // onClick={() => navigate('/', { replace: true })}
+                onClick={() => setShowWelcomeModal(true)}
+                buttonType="primary"
+                className="w-1/2 px-4 py-2"
+                disabled={nextDisabled}
+              >
+                Finish
+              </Button>
+            )
+            : (
+              <Button
+                onClick={() => handleNext()}
+                buttonType="primary"
+                className="w-1/2 px-4 py-2"
+                disabled={nextDisabled || isFetching}
+                loading={isFetching}
+              >
+                Next
+              </Button>
+            )}
+        </div>
       </div>
-      <div className="flex justify-between font-semibold">
-        <Button onClick={() => navigate(-1)} buttonType="primary" className="mr-6 w-1/2 py-2" disabled={prevDisabled}>
-          Back
-        </Button>
-        {finish
-          ? (
-            <Button
-              onClick={() => navigate('/', { replace: true })}
-              buttonType="primary"
-              className="w-1/2 px-4 py-2"
-              disabled={nextDisabled}
-            >
-              Finish
-            </Button>
-          )
-          : (
-            <Button
-              onClick={() => handleNext()}
-              buttonType="primary"
-              className="w-1/2 px-4 py-2"
-              disabled={nextDisabled || isFetching}
-              loading={isFetching}
-            >
-              Next
-            </Button>
-          )}
-      </div>
-    </div>
+      <WelcomeModal onClose={() => navigate('/', { replace: true })} show={showWelcomeModal} />
+    </>
   );
 }
